requirements.txt
pydantic>=2.0
requests
pandas
pyarrow
python-dotenv
pytest
pyspark
delta-spark